+++ src/midi/WindowsMidiServicesTypeReferenceProbe.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <QString>
+
+class WindowsMidiServicesTypeReferenceProbe {
+public:
+    static QString typeReferenceExperimentStatus();
+};